requirements.txt
openai
flask
python-dotenv
requests
azure-identity
ipykernel